Refactor into a single file, clean up code
