--- ansible_base/resource_registry/workload_identity_types.py
@@ -1,31 +1,0 @@
-"""Type definitions for workload identity client."""
-
-from typing import NamedTuple, Optional
-
-
-class WorkloadIdentityTokenRequest(NamedTuple):
-    """
-    Request body for workload identity token endpoint.
-    """
-
-    claims: dict
-    """Dictionary of claims to include in the workload identity token."""
-
-    scope: str
-    """Token scope string (e.g., 'read', 'write', 'read write')."""
-
-
-class WorkloadIdentityTokenResponse(NamedTuple):
-    """Response from workload identity token endpoint."""
-
-    access_token: str
-    """The JWT access token."""
-
-    token_type: Optional[str] = "Bearer"
-    """Token type, typically 'Bearer'."""
-
-    expires_in: Optional[int] = None
-    """Token expiration time in seconds."""
-
-    scope: Optional[str] = None
-    """The scope of the token."""
